+++ BetterChests/Enums/SortBy.cs
@@ -1,0 +1,19 @@
+﻿namespace StardewMods.BetterChests.Enums;
+
+/// <summary>
+///     Sorting type for items.
+/// </summary>
+internal enum SortBy
+{
+    /// <summary>Default sorting.</summary>
+    Default = 0,
+
+    /// <summary>Sort by type.</summary>
+    Type = 1,
+
+    /// <summary>Sort by quality.</summary>
+    Quality = 2,
+
+    /// <summary>Sort by quantity.</summary>
+    Quantity = 3,
+}