+++ BetterChests/Interfaces/ManagedObjects/IManagedStorage.cs
@@ -7,6 +7,11 @@ using StardewValley;
 internal interface IManagedStorage : IManagedObject, IStorageContainer
 {
     /// <summary>
+    ///     Gets or sets a value indicating whether organize chest should sort in reverse order.
+    /// </summary>
+    public bool OrganizeChestOrderByDescending { get; set; }
+
+    /// <summary>
     ///     Attempt to stash an item into the storage based on categorization and existing items.
     /// </summary>
     /// <param name="item">The item to stash.</param>